Client counter animation on scroll update
