+++ js/main-home.js
@@ -15,7 +15,6 @@ import { renderMySkills } from "./components/mySkills/renderMySkills.js";
 /* client counter */
 import { clientCounterData } from './data/clientCounterData.js';
 import { renderClientCounter } from './components/clientCounter/renderClientCounter.js';
-import { clientCounterAnimation } from './components/clientCounter/clientCounterAnimation.js';
 import { counterAnimationOnScroll } from './components/clientCounter/counterAnimationOnScroll.js';
 /* services */
 import { servicesData } from "./data/servicesData.js";
@@ -46,7 +45,6 @@ renderMySkills('#skillsRight', skillsData);
 /* resume */
 /* client counter */
 renderClientCounter('#clientCounterBlock', clientCounterData);
-clientCounterAnimation();
 counterAnimationOnScroll();
 /* services */
 renderMyServices('#services-block', servicesData);
